requirements.txt
Django==3.2.5
djangorestframework==3.11.0
psycopg2-binary>=2.8
openpyxl
xlrd